Add WindowsProfile fields
Add fields to specify an admin username and password for Windows node
pool types, which are required in order to set OSType to "Windows".
Rename adminUsername to linuxAdminUsername to differentiate it. Add a
new Windows-specific CR example, showing that networkPlugin must be
"azure" and that a virtualNetwork and subnet are required.