Use same SLES image as Rancher
In addition to using the SLES image for the aks-operator image, the
dapper image is also switched to using SLES.

A side effect of this is that docker will no longer be installed in the
dapper image. Instead, a new pipeline step is added to build the image
on pull requests.