Wait for non-Rancher initiated updates to complete
If a user updates the AKS cluster in AKS console, a transitory state may
be synced. Once Rancher sees this middle state, it will try to reconcile
back to this undesired state after the update completes in AKS.

With this change, aks-operator recognizes when an update was non
initiated by Rancher. It will then ignore the update until the AKS
cluster is in a "Succeeded" state.

Co-authored-by: Ricardo Weir <ricardo@rancher.com>