Lookup credential tenant ID if not provided
Use FindTenantID from rancher/machine to look up a tenant ID from a
subscription ID if none is provided in the cloud credential. This makes
providing the tenant ID optional. The tenant ID is not stored, it is
looked up every time a client is created, which is every reconcile loop.