Store fetched Azure tenant ID in secret annotation
In order to avoid making repetitive requests to the Azure cloud
authorization endpoint to retrieve the tenant ID, store the fetched
tenant ID in an annotation on the cloud credential secret. The
annotation expires after an hour and will be re-fetched. Although it is
not likely to change hour to hour, in the event that it does change we
want the credential to be usable within a reasonable delay.